fix(ui): address review feedback
- Fix loadData() overwriting activeProjectId on every call, which reset
  the user's project dropdown selection. Now initialized once via
  LaunchedEffect(Unit).
- Extract shared DateUtils (formatEpochDate, dateToEpochRange, isLeapYear)
  to core/util/DateUtils.kt, eliminating 3x duplication across
  SemanticCompressor, DigestGenerator, and ArticlesScreen.
- Clean up fully qualified Koin import in ArticlesScreen.

Refs #33
